# Conflicts:
#	docs/categories/index.html
#	docs/index.html
#	docs/posts/26/index.html
#	docs/posts/28/index.html
#	docs/posts/37/index.html
#	docs/posts/40/index.html
#	docs/posts/41/index.html
#	docs/posts/43/index.html
#	docs/posts/44/index.html
#	docs/posts/54/index.html
#	docs/posts/55/index.html
#	docs/posts/72/index.html
#	docs/posts/index.html
#	docs/tags/curl/index.html
#	docs/tags/docker/index.html
#	docs/tags/dynamodb/index.html
#	docs/tags/gas/index.html
#	docs/tags/homebrew/index.html
#	docs/tags/index.html
#	docs/tags/macm1/index.html
#	docs/tags/mysql5.6/index.html
#	docs/tags/php/index.html
#	docs/tags/ruby/index.html
#	docs/tags/yum/index.html
#	hugo/assets/sass/main.scss
#	hugo/docs/categories/index.html
#	hugo/docs/index.html
#	hugo/docs/posts/26/index.html
#	hugo/docs/posts/28/index.html
#	hugo/docs/posts/37/index.html
#	hugo/docs/posts/40/index.html
#	hugo/docs/posts/41/index.html
#	hugo/docs/posts/43/index.html
#	hugo/docs/posts/44/index.html
#	hugo/docs/posts/54/index.html
#	hugo/docs/posts/55/index.html
#	hugo/docs/posts/72/index.html
#	hugo/docs/posts/index.html
#	hugo/docs/tags/curl/index.html
#	hugo/docs/tags/docker/index.html
#	hugo/docs/tags/dynamodb/index.html
#	hugo/docs/tags/gas/index.html
#	hugo/docs/tags/homebrew/index.html
#	hugo/docs/tags/index.html
#	hugo/docs/tags/macm1/index.html
#	hugo/docs/tags/mysql5.6/index.html
#	hugo/docs/tags/php/index.html
#	hugo/docs/tags/ruby/index.html
#	hugo/docs/tags/yum/index.html
#	hugo/layouts/_default/single.html
#	hugo/layouts/index.html
#	hugo/resources/_gen/assets/scss/sass/main.scss_48b060fe05b0a273d182ef83c0605941.content
#	hugo/resources/_gen/assets/scss/sass/main.scss_48b060fe05b0a273d182ef83c0605941.json

+++ hugo/copy-issue.rb
@@ -70,4 +70,6 @@ File.open(file_path, 'a') do |f|
   end
 end
 
-File.delete("#{file_path}-e")
+if File.exist?("#{file_path}-e")
+  File.delete("#{file_path}-e")
+end
